Rename errorboundary directory to error
The directory contains error related components now not only for the
error boundary.
