+++ gsa/CMakeLists.txt
@@ -245,8 +245,9 @@ set (GSA_JS_SRC_FILES
      ${GSA_SRC_DIR}/src/web/components/dialog/scrollablecontent.js
      ${GSA_SRC_DIR}/src/web/components/dialog/title.js
      ${GSA_SRC_DIR}/src/web/components/dialog/twobuttonfooter.js
-     ${GSA_SRC_DIR}/src/web/components/errorboundary/errorboundary.js
-     ${GSA_SRC_DIR}/src/web/components/errorboundary/errorcontainer.js
+     ${GSA_SRC_DIR}/src/web/components/error/errorboundary.js
+     ${GSA_SRC_DIR}/src/web/components/error/errorcontainer.js
+     ${GSA_SRC_DIR}/src/web/components/error/errormessage.js
      ${GSA_SRC_DIR}/src/web/components/folding/folding.js
      ${GSA_SRC_DIR}/src/web/components/footnote/footnote.js
      ${GSA_SRC_DIR}/src/web/components/form/button.js
