+++ gsa/src/web/components/error/errorcontainer.js
