+++ gsa/src/web/entity/page.js
@@ -26,7 +26,7 @@ import PropTypes from 'web/utils/proptypes';
 
 import Toolbar from 'web/components/bar/toolbar';
 
-import ErrorMessage from 'web/components/errorboundary/errormessage';
+import ErrorMessage from 'web/components/error/errormessage';
 
 import Layout from 'web/components/layout/layout';
